add variable 'pkgs' to lib file and to the record of all pkgs
Thank to recursive merging and lazy evaluation, the record with all packages can contain a reference to itself.
This makes it possible to access other packages like so:
{
	pkgs,
	myPacakge={
		biPkgs=[ pkgs.foo, pkgs.bar ]
	}
}

The library is now defined in lib2.ncl (lib.ncl is a combat thing) like so:

{
	pkgs,
	lib={..}
}
whick means that default dependencies can be added bu using pkgs in the library.